Helpers fixed and Timers enhanced
The helper draggables now correctly float above the over page elements.
The timer has been enhanced such that when a command is made, the timer
is stopped, the pictures updated, the timer restarted, and a 5 second
timeout kicked off to ensure the first refresh happens on time and soon
enough.